Added maven. Redid some of the FileTransferProtocol code for improved readability
