--- client/src/Helpers/CleanupCallback.java
@@ -1,6 +1,0 @@
-package Helpers;
-
-public interface CleanupCallback {
-    void performCleanup();
-}
-
