--- client/src/Managers/SessionManager.java
@@ -1,30 +1,0 @@
-package Managers;
-
-import java.util.Map;
-import java.util.HashMap;
-
-public class SessionManager {
-
-    private Map<String, String> publicKeys;
-    private boolean loggedIn = false;
-
-    public SessionManager() {
-        publicKeys = new HashMap<>();
-    }
-
-    public Map<String, String> getPublicKeys() {
-        return publicKeys;
-    }
-
-    public void addPublicKey(String username, String publicKey) {
-        this.publicKeys.put(username, publicKey);
-    }
-
-    public boolean isLoggedIn() {
-        return loggedIn;
-    }
-
-    public void setLoggedIn(boolean loggedIn) {
-        this.loggedIn = loggedIn;
-    }
-}
