+++ client/src/main/java/Helpers/CleanupCallback.java
@@ -1,0 +1,6 @@
+package main.java.Helpers;
+
+public interface CleanupCallback {
+    void performCleanup();
+}
+
